// src/modules/blogs/types.ts
export namespace BlogsData {
  export interface Blogs {
    schema: any;
    articleSection: string;
    date: string;
    title: string;
    url: string;
    id?: number;
    minutes: string;
    link: string;
    slug: string;
    twitter_img: string;
    author: string;
    avatarUrl: string;
    description: string;
  }
}